Fixed the bug of invisible sprites
If the sprite is set to scale 0 in the first frame of animation, Unity can sometimes decide to not draw it at all, even if the scale changes over the animation. Deactivate the GameObject completely to be able to use a scale of e.g. 0.01, so the sprite is initialized anyways.